update from android

+++ database/migrations/2020_09_03_061416_create_courts_table.php
@@ -18,8 +18,8 @@ class CreateCourtsTable extends Migration
             $table->string('name');
             $table->string('address');
             $table->string('phone');
-            $table->bigInteger('lat');
-            $table->bigInteger('lng');
+            $table->double('lat');
+            $table->double('lng');
             $table->timestamps();
         });
     }
